More old code cleanup
