+++ src/encoded/static/components/index.js
@@ -11,7 +11,6 @@ require('./lib');
 require('./util');
 require('./viz/utilities');
 require('./app');
-require('./collection');
 require('./footer');
 require('./globals');
 require('./static-pages/static-page-base');
@@ -34,7 +33,7 @@ require('./item-pages/FileMicroscopyView');
 require('./item-pages/BiosampleView');
 require('./item-pages/BiosourceView');
 require('./item-pages/ProtocolView');
-require('./statuslabel');
+//require('./statuslabel');
 require('./navigation');
 require('./submission/submission-view');
 require('./submission/submission-fields');
